requirement.txt
aniso8601==7.0.0
Django>=2.0,<3.0
django-cors-headers==3.0.2
django-graphql-jwt==0.2.2
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
promise==2.2.1
PyJWT==1.7.1
pytz==2019.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
typing==3.7.4.1